Backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
email_validator==2.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
Pygments==2.18.0
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.2
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
watchfiles==1.0.3
websockets==14.1
Werkzeug==3.1.3